Mostly working parsing
There's a few bugs in Reflect.parse it seems, where it reports the wrong column
values. There's a special case for FunctionDeclaration as the body includes the
start `{` but not the end `}`.

It will choke on `let foo = () => {}` though, as declaration.init.loc includes
the equal sign for some reason.